requirements.txt
Flask
pandas
matplotlib
seaborn
scikit-learn
requests
pyserial